HarvestSync-Algorithm_API/requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0

# Logging ---------------------------------------------------------------------
loguru~=0.5.3

# Plotting --------------------------------------------------------------------
pandas>=1.1.4

# API ----------------------------------------------------------------
uvicorn~=0.23.2
fastapi~=0.101.1
pydantic~=2.1.1

torch~=2.0.1
seaborn~=0.12.2
ipython~=8.8.0
torchvision~=0.15.2
tqdm~=4.65.0
psutil~=5.9.0